Fixed issue with discovery not scanning sources.
