+++ src/Beta.TestAdapter/VsTestAdapter.cs
@@ -1,44 +1,37 @@
-//using Beta.TestAdapter.Models;
-//using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
-//using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
-//using System.Diagnostics;
+using Beta.TestAdapter.Models;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 
-//namespace Beta.TestAdapter;
+namespace Beta.TestAdapter;
 
-//public class VsTestAdapter
-//{
-//    public VsTestAdapter()
-//    {
-//        Debugger.Launch();
-//        Debugger.Break();
-//    }
+public class VsTestAdapter
+{
+    /// <summary>
+    ///     Gets the internal logger.
+    /// </summary>
+    protected ITestLogger Logger { get; private set; } = TestLogger.Null;
 
-//    /// <summary>
-//    ///     Gets the internal logger.
-//    /// </summary>
-//    protected ITestLogger Logger { get; private set; } = TestLogger.Null;
+    /// <summary>
+    ///     Gets the run settings for the adapter.
+    /// </summary>
+    protected RunSettings Settings { get; private set; } = new();
 
-//    /// <summary>
-//    ///     Gets the run settings for the adapter.
-//    /// </summary>
-//    protected RunSettings Settings { get; private set; } = new();
+    /// <summary>
+    ///     Resets the adapter with the given context.
+    /// </summary>
+    /// <param name="context">The context to reset with.</param>
+    /// <param name="logger">The current logger.</param>
+    protected void Reset(IDiscoveryContext? context, IMessageLogger? logger)
+    {
+        Settings = RunSettings.Parse(context?.RunSettings?.SettingsXml);
+        Logger = new TestLogger(logger);
+    }
 
-//    /// <summary>
-//    ///     Resets the adapter with the given context.
-//    /// </summary>
-//    /// <param name="context">The context to reset with.</param>
-//    /// <param name="logger">The current logger.</param>
-//    protected void Reset(IDiscoveryContext? context, IMessageLogger? logger)
-//    {
-//        Settings = RunSettings.Parse(context?.RunSettings?.SettingsXml);
-//        Logger = new TestLogger(logger);
-//    }
-
-//    /// <summary>
-//    ///     Prints the current banner.
-//    /// </summary>
-//    protected void PrintBanner()
-//    {
-//        Logger.Info($"Target Framework Version: {Settings.Configuration.TargetFrameworkVersion}");
-//    }
-//}
+    /// <summary>
+    ///     Prints the current banner.
+    /// </summary>
+    protected void PrintBanner()
+    {
+        Logger.Info($"Target Framework Version: {Settings.Configuration.TargetFrameworkVersion}");
+    }
+}
